backend/requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
pymysql==1.1.0
alembic==1.12.1

# Authentication
PyJWT==2.8.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0

# Data Processing
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
scipy==1.11.4

# API Documentation
python-multipart==0.0.6

# Caching
redis==5.0.1

# File handling
openpyxl==3.11.0
xlrd==2.0.1

# Utilities
email-validator==2.1.0
requests==2.31.0
httpx==0.25.2

# Development
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1
black==23.12.0
flake8==6.1.0
mypy==1.7.0
